[KIECLOUD-277] - JMS Parameters available for user to configure but not applied on kieserver template.
While adding the two missing parameters to the template was identified if, when a default value of
the JMS boolean parameter is true, if this parameter is empty or false, the default will be true.
To fix that, boolean pointers was used, in that way when the param is empty or false, it will be
correctly set.

Signed-off-by: Filippe Spolti <fspolti@redhat.com>